+++ src/test/groovy/ru/vyarus/guice/persist/orient/finder/internal/AbstractFinderDefinitionTest.groovy
@@ -3,8 +3,8 @@ package ru.vyarus.guice.persist.orient.finder.internal
 import com.google.inject.Inject
 import ru.vyarus.guice.persist.orient.AbstractTest
 import ru.vyarus.guice.persist.orient.db.transaction.template.SpecificTxAction
-import ru.vyarus.guice.persist.orient.finder.internal.generics.FinderGenericsFactory
-import ru.vyarus.guice.persist.orient.finder.internal.generics.GenericsDescriptor
+import ru.vyarus.java.generics.resolver.GenericsResolver
+import ru.vyarus.java.generics.resolver.context.GenericsContext
 
 import java.lang.reflect.Method
 
@@ -16,15 +16,13 @@ abstract class AbstractFinderDefinitionTest extends AbstractTest {
 
     @Inject
     FinderDescriptorFactory factory
-    @Inject
-    FinderGenericsFactory genericsFactory
 
     FinderDescriptor lookup(Method method) {
         lookup(method.getDeclaringClass(), method)
     }
 
     FinderDescriptor lookup(Class iface, Method method) {
-        GenericsDescriptor generics = genericsFactory.create(iface);
+        GenericsContext generics = GenericsResolver.resolve(iface);
         template.doInTransaction({
             return factory.create(method, generics)
         } as SpecificTxAction)