+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/delegate/Finder.groovy
@@ -6,7 +6,7 @@ import ru.vyarus.guice.persist.orient.support.model.Model
  * @author Vyacheslav Rusakov 
  * @since 22.10.2014
  */
-public interface Finder extends Mixin<Model>, ComplexFinder<Model> {
+public interface Finder extends Mixin<Model>, ComplexFinder<Model>, Errors {
 
     void countIndirect();