+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/generics/ComplexGenerics.groovy
@@ -1,9 +1,0 @@
-package ru.vyarus.guice.persist.orient.support.finder.generics
-
-/**
- * @author Vyacheslav Rusakov 
- * @since 19.10.2014
- */
-public interface ComplexGenerics<T, K extends Collection<T>> {
-
-}