+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/generics/Lvl2Base3.groovy
@@ -1,9 +1,0 @@
-package ru.vyarus.guice.persist.orient.support.finder.generics
-
-/**
- * @author Vyacheslav Rusakov 
- * @since 16.10.2014
- */
-public interface Lvl2Base3<R> {
-
-}