+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/generics/Lvl2BeanBase.groovy
@@ -1,8 +1,0 @@
-package ru.vyarus.guice.persist.orient.support.finder.generics
-
-/**
- * @author Vyacheslav Rusakov 
- * @since 18.10.2014
- */
-class Lvl2BeanBase<I> {
-}